Cleanup formatting and tooltip code
Removes duplication in WIP stuff and moves text/tooltip/json generation
into a new `send_output` function so it's a little easier to parse.

Also removes the `function` from the rest of the functions since `usage`
doesn't have it so it makes more sense that the rest also don't get it.